script/windows-installer: show deregistration text
When uninstalling Git LFS on Windows, as of commit
6de1bcea13867d945ac922d9a730c2cbb1c6d182 [formerly cbd52be073a340808253208ee4430b6f6451c271] we now use the
term "deregister" in part of the message we display when
a Git exectuable is found in an unconventional location.

Since we also use the term "registered" in another part
of that message, we change it to switch to the term
"deregistered" in the uninstallation case.


Former-commit-id: 5a9f978db582604eb556a18511e0d8bb45e67165