lfs: don't write hooks when they haven't changed
Some people store their hooks inside their repository.  This isn't a
secure approach, but in some cases it can be acceptable if all users are
trusted.

In such a case, if Git LFS always rewrites a hook even if it hasn't
changed, it can result in the file being marked as modified by `git
status`, since `git status` can invoke Git LFS, which in turn invokes
Git LFS, which updates the hooks.  Thus, the hooks will always appear
modified.

Let's solve this by not writing the hooks if they haven't changed.  We
don't include tests here because the only way to determine if changes
have occurred is the mtime, and that's difficult to effectively test in
the testsuite in a portable way without introducing sleeps, which would
make our tests slow.


Former-commit-id: e0551f8be721818f3412f5e5bf61d5a42ac850eb