Makefile: remove legacy trimpath code
Before Go 1.13, it was possible to trim paths with GC flags and ASM
flags.  In Go 1.13, it became possible to simply specify -trimpath, at
which point all the old behaviour became unnecessary.

Since we require Go 1.13 in our code due to the use of the errors
package, let's simplify our code by always using the -trimpath flag and
omitting the old forms.  The old forms additionally were broken when
using a path with a space, which could happen on Windows if the current
user's username contained a space (which is not uncommon).


Former-commit-id: 7809d689d24fb564ab6f53316f755e1abad4e9bc