t/Makefile: build test helpers with file extension
On Windows, it is convenient to have the .exe file extension
added to our compiled test helpers when running tests manually,
so we use the X variable to do so in the same manner as in the
primary Makefile.

Also, we add a test-commands Makefile target which is convenient
when re-building just the test helpers.


Former-commit-id: 5f7bd242a949c3087e49873d3bc2cc2cc1b95547