t/t-prune.sh: edit comments and drop unused prune
We revise some of the comments in the "prune keep unpushed"
test to clarify their meaning before we add additional checks
to this test.

We also drop an unused "git lfs prune --dry-run" invocation
since we are not testing its output and therefore it serves no
purpose in the test.


Former-commit-id: d85d28d193bb88e97daf4f43be0538819d3a21f4