Makefile: put all archive contents in a directory
Right now, extracting one of our archives (except for the source
archives) causes all files to be extracted in the current directory.
This is commonly referred to as a tarbomb, and is generally unwelcome.

Instead, let's create a top-level directory in each of our archives
named the same way as we currently name our source tarballs: the string
"git-lfs-" and then a version number without the initial "v".  This is
the same algorithm GitHub would use to create its autogenerated
archives.

Because we can no longer use zip's -j option to strip off path
components (since we now want to preserve at least some of them), using
zip becomes tricky.  Instead, let's use the BSD (libarchive) tar, which
is natively available on Windows, macOS, and FreeBSD as the system tar
implementation, and which can be easily installed on all Linux distros
under the name "bsdtar".  This program supports tar files, zip files,
and a wide variety of other formats that we don't need here and is ideal
for this usage.  However, it doesn't support updating zip files in
place, so we need to extract and rebuild the archives instead.


Former-commit-id: d7278e35465c48cf3fb72df8ac4cabe0ca003790