.github/workflows: install bsdtar on Linux
On macOS and Windows, the default tar implementation is the BSD
(libarchive) tar, which handles various formats.  We need this tar
implementation to handle the creation and modification of zip files,
which GNU tar doesn't support.  As a result, install the
libarchive-tools package on Linux as well so we can make use of it.


Former-commit-id: 31003a82e7711f8696c308f238f21db666f09d58