git: don't panic on pkt-line without equals
In some cases, we can end up with a version of Git that contain a
filter-process pkt-line without an equals.  (What exactly these contents
are is not known.)  This causes us to panic when we split, since `v[1]`
doesn't exist in the array.  Let's fix this by simply ignoring the
header if it doesn't contain an equals since we wouldn't care about its
value anyway.


Former-commit-id: 2663a30319ad5eafd9ffbd831b815a8511131ab4