workflows: use ronn-ng
Recently, we changed to use the windows-2019 target because ronn fails
to build on newer versions.  Unfortunately, it also fails to build on
that version due to updates in GitHub Actions.

Since ronn itself is dead upstream, let's switch to ronn-ng, just like
Debian has done.  This provides an identically named binary but is a
supported codebase that is more likely to work now and in the future.


Former-commit-id: 912b607acb162ea324d0ff00fea1a1f47657e1d4