Makefile: specify path to bsdtar on Windows
On Windows, the tar shipped with the OS is a BSD (libarchive) tar.
However, under Git for Windows, the tar command is a GNU tar, which is
also the case in GitHub Actions.  As a result, we need to explicitly
specify the path for bsdtar in a different way than for Unix systems,
where we can just typically fall back to "bsdtar", so let's specify the
full path.


Former-commit-id: 343dbed73e85626e4ac24e6d5c67f089931e25ed