Support for multiple remotes to solve issue #3835
Before this change, LFS was only respecting remotes which were directly
tracked by a branch, or, in a detached state, using a heuristic.
Sometimes this failed as reported in issue #3835.
This patch allows to use the tree-ish to retrieve a corresonding remote.

Thanks to Stan Hu, who provided inspiration to fix this bug.


Former-commit-id: 71237a1ef4866c5cba140f20a0a7497712d0ec2f