t: fix standalone tests on Windows
In our shell script tests, we can end up with paths that start with
things like "/tmp", which on Windows is implicitly rewritten into a real
Windows path by the shell.  However, when using the `urlify` function,
that doesn't happen and therefore Git LFS gets a path it cannot handle.

We used to paper over this by calling `cygpath -w` on all the paths we
were given, but the version of `cygpath` shipped with Git for Windows is
subtly broken and we don't want to rely on it anymore, so instead make
our `urlify` helper convert to a canonical native path and then prefix
with a slash if need be.

Similarly, use the `urlify` helper in the multiple remotes test to make
sure that the paths are rewritten properly.


Former-commit-id: b3955340247f9ecfca4189bffa00e44199d4be1e