ssh: disable concurrent transfers if no multiplexing
Right now, we try to spawn multiple SSH connections using ControlMaster
when making multiple requests.  However, if multiplexing is not enabled,
we can spawn multiple actual connections, which can be expensive and
require lots of authentication requests.

Instead, let's make sure we don't enable multiple connections if
multiplexing is not enabled, since this may cause the user to be
prompted for multiple SSH key connection attempts (say, if they're
using a security key) and is substantially more heavyweight than simply
spawning a new process over the same connection.

Since the code has different behaviour if `XDG_RUNTIME_DIR` is set, make
sure to unset it for the tests so that we always try to create a
temporary directory and don't otherwise fail because that environment
variable points somewhere unexpected.


Former-commit-id: b44cbe40ef4163744ea4f2b30b738924a969e962