ssh: don't leak resources when falling back to legacy protocol
When we attempt a connection with the pure SSH protocol but fail, we
fail to close the file descriptors for the pipes or reap the process
which has exited.  This can cause substantial resource consumption and
run the user out of file descriptors or processes.

Since that's not very nice, let's be sure to explicitly close the pipes
and wait for the process if this occurs, which it probably will quite
frequently, since few people are using the new pure SSH protocol.


Former-commit-id: bc06dc22e4c8d51c469f154954895643fc84d983